requirements.txt
asyncio==4.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
greenlet==3.2.4
idna==3.11
playwright==1.55.0
pyee==13.0.0
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
pika==1.3.2
redis==5.0.0
python-dotenv==1.2.1
